uint8_t m_frameIndex.

+++ IrradianceMap/IrradianceMap.h
@@ -68,7 +68,7 @@ private:
 	bool		m_typedUAV;
 
 	// Synchronization objects.
-	uint32_t	m_frameIndex;
+	uint8_t		m_frameIndex;
 	HANDLE		m_fenceEvent;
 	XUSG::Fence	m_fence;
 	uint64_t	m_fenceValues[FrameCount];
